git subrepo pull kigs
subrepo:
  subdir:   "kigs"
  merged:   "5b9b5ca370"
upstream:
  origin:   "https://github.com/assoria/kigs.git"
  branch:   "master"
  commit:   "5b9b5ca370"
git-subrepo:
  version:  "0.4.1"
  origin:   "???"
  commit:   "???"

Former-commit-id: 93a41004256533586ce1fd0d8ed46f7cac40ea54
Former-commit-id: b26ed3510c51b65d7ce4d82f9f24fdbc3a3ebc33

+++ framework/TinyImage/Sources/JPEGClass.cpp
@@ -244,7 +244,7 @@ std::vector<u8> RGB24ToJpeg(unsigned char* image_buffer, int image_width, int im
 		* requires it in order to write binary files.
 		*/
 	u8* outbuffer=nullptr;
-	unsigned long outsize=0;
+	size_t outsize=0;
 	jpeg_mem_dest(&cinfo, &outbuffer, &outsize);
 	/* Step 3: set parameters for compression */
